files-view: Stop doing ranged model searches
Ranged model searching was a minor optimization meant to make the next
binary search smaller.

But we must drop it in order to apply an incompatible fix to a major
flaw in the binary search implementation in the next commit. As such,
there is no reason to sort list of items in advance either.

This partially reverts commit bbd7aa7798ab8b710fd9cfe8eb2008663b43a948.