view-model: Account for tree sorting in find()
The binary search assumes relies on the sort order. This works fine as
long as we don't have expanded folders, where the comparing item for
sorting is not enough: we need to compare their relative positions in
the tree hierarchy.

When folders are expanded as a tree, we may fail to find and crash.

Let's do a tree-adjusted comparison where we find the common parent and
then compare recursively until a difference is found.