files-view: Manage |view.visible-columns
The action group is owned by NautilusFilesView and the only reason why
NautilusListView can add this action is because it's a subclass.

But in the future we want it to be a child, and possible always present
even not shown, so only NautilusFilesView can decide when this action
should be shown or not.

Also remove dead dupicate menu definition from progress-indicator.ui,
where this action was referenced for no reason.